requirements.txt
# MindCanvas Backend requirements
# Usage: pip install -r requirements.txt

# Core Dependencies -----------------------------------------------------------
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2

# Web Application -------------------------------------------------------------
Flask>=2.0.0
Werkzeug>=2.0.0
flask-cors>=3.0.0

# YOLOv5 Dependencies ---------------------------------------------------------
yolov5>=6.0.0
PyYAML>=5.3.1
requests>=2.25.0
scipy>=1.4.1
tqdm>=4.64.0
matplotlib>=3.2.2
pandas>=1.1.4
seaborn>=0.11.0

# Additional Dependencies -----------------------------------------------------
python-dotenv>=0.19.0
